+++ StripeFinancialConnections/StripeFinancialConnections/Source/API Bindings/Models/FinancialConnectionsDataAccessNotice.swift
@@ -8,14 +8,20 @@
 import Foundation
 
 struct FinancialConnectionsDataAccessNotice: Decodable {
+    let icon: FinancialConnectionsImage?
     let title: String
+    let connectedAccountNotice: ConnectedAccountNotice?
     let subtitle: String?
     let body: Body
-    let connectedAccountNotice: String?
-    let learnMore: String
+    let disclaimer: String?
     let cta: String
 
     struct Body: Decodable {
         let bullets: [FinancialConnectionsBulletPoint]
+    }
+
+    struct ConnectedAccountNotice: Decodable {
+        let subtitle: String
+        let body: Body
     }
 }