+++ StripeFinancialConnections/StripeFinancialConnections/Source/Native/Shared/PaneLayoutView.swift
@@ -24,7 +24,7 @@ final class PaneLayoutView {
     init(contentView: UIView, footerView: UIView?) {
         self.scrollViewContentView = contentView
 
-        let scrollView = UIScrollView()
+        let scrollView = AutomaticShadowScrollView()
         self.scrollView = scrollView
         scrollView.addAndPinSubview(contentView)
 
@@ -42,10 +42,75 @@ final class PaneLayoutView {
     }
 
     func addTo(view: UIView) {
-        // this function encapsulates an error-prone sequence where we
+        // This function encapsulates an error-prone sequence where we
         // must add `paneLayoutView` (and all it's subviews) to the `view`
         // BEFORE we can add a constraint for `UIScrollView` content
         view.addAndPinSubviewToSafeArea(paneLayoutView)
         scrollViewContentView?.widthAnchor.constraint(equalTo: view.safeAreaLayoutGuide.widthAnchor).isActive = true
+
+        // Fit the scroll view height to be the size of the
+        // scroll view contents
+        //
+        // For exampple, this is needed for `SheetViewController`
+        // to automatically re-size the sheet to the size of contents
+        let scrollViewHeightConstraint = scrollView.heightAnchor.constraint(
+            equalTo: scrollView.contentLayoutGuide.heightAnchor)
+        scrollViewHeightConstraint.priority = .fittingSizeLevel
+        scrollViewHeightConstraint.isActive = true
+    }
+
+    func createView() -> UIView {
+        let containerView = UIView()
+        addTo(view: containerView)
+        return containerView
+    }
+}
+
+// Automatically adds a shadow to the bottom
+// if the content is scrollable
+private class AutomaticShadowScrollView: UIScrollView {
+
+    private var shadowView: UIView?
+
+    override func layoutSubviews() {
+        super.layoutSubviews()
+
+        let canScroll = contentSize.height > bounds.height
+        if canScroll && shadowView == nil {
+            let shadowView = UIView()
+            self.shadowView = shadowView
+            shadowView.layer.shadowColor = UIColor.textDefault.cgColor
+            shadowView.layer.shadowOpacity = 0.77
+            shadowView.layer.shadowOffset = CGSize(width: 0, height: -4)
+            shadowView.layer.shadowRadius = 10
+            // if the background color is clear, iOS will
+            // not draw a shadow
+            shadowView.backgroundColor = UIColor.customBackgroundColor
+            addSubview(shadowView)
+        } else if !canScroll {
+            shadowView?.removeFromSuperview()
+            shadowView = nil
+        }
+
+        if let shadowView {
+            // smaller shadow width "smoothens" the shadow 
+            // around the leading/trailing edges
+            let x = Constants.Layout.defaultHorizontalMargin / 2
+            // move the `shadowView` to keep being at the bottom of visible bounds
+            shadowView.frame = CGRect(
+                x: x,
+                y: contentOffset.y + bounds.height,
+                width: bounds.width - (2 * x),
+                height: 1
+            )
+
+            // slowly fade the `shadowView` as user scrolls to bottom
+            //
+            // the fade will only activate when we reach `startFadingDistanceToBottom`
+            let distanceToBottom = contentSize.height - (contentOffset.y + bounds.size.height)
+            let startFadingDistanceToBottom: CGFloat = 24
+            let remainingFadeDistance = max(0, min(startFadingDistanceToBottom, distanceToBottom))
+            shadowView.alpha = remainingFadeDistance / startFadingDistanceToBottom
+        }
     }
 }